+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Add network constants for Base Mainnet, Ethereum Holesky, and Ethereum Mainnet
+
 ## [0.0.13] - 2024-07-30
 
 ### Added